% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_props_commercial_fn}
\alias{age_props_commercial_fn}
\title{The filename for the survey ge proportions by year data file}
\format{
A character string
}
\usage{
age_props_commercial_fn
}
\description{
The filename for the survey ge proportions by year data file
}
\keyword{datasets}
