% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_props_fn}
\alias{age_props_fn}
\title{The filename for the Freezer trawler age proportions by year data file}
\format{
A character string

A character string
}
\usage{
age_props_fn

age_props_fn
}
\description{
The filename for the Freezer trawler age proportions by year data file

The filename for the Shoreside age proportions by year data file
}
\keyword{datasets}
