% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catch_targets_df}
\alias{catch_targets_df}
\title{A data frame containing information on the actual catch, and Total
Allowable Catch by year}
\format{
A data frame with 3 variables:
\describe{
\item{Year}{Year of catch/TAC}
\item{Catch}{Actual catch take for the year in tonnes}
\item{TAC}{Total allowable catch for the year}
}
}
\usage{
catch_targets_df
}
\description{
A data frame containing information on the actual catch, and Total
Allowable Catch by year
}
\keyword{datasets}
