% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check-catch-levels.R
\name{check_catch_levels}
\alias{check_catch_levels}
\title{Check that catch levels has been run}
\usage{
check_catch_levels(
  model_path = NULL,
  forecast_yrs = get_assess_yr():(get_assess_yr() + 3),
  ...
)
}
\arguments{
\item{model_path}{The directory the models are located in}

\item{forecast_yrs}{A vector of forecast years}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
Logical. \code{TRUE} if the catch levels appears to have been run
successfully and \code{FALSE} otherwise
}
\description{
Check that catch levels has been run
}
\details{
Ensure that the \code{\link[=run_catch_levels]{run_catch_levels()}} function has been run successfully
by looking in to the catch levels directory and checking values in the
forecast files
}
