% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create-age-proportions-csv.R
\name{create_age_proportions_csv}
\alias{create_age_proportions_csv}
\title{Create an age proportion CSV data file from the sample data}
\usage{
create_age_proportions_csv(
  d,
  min_date = as.Date("1979-01-01"),
  raw_counts = FALSE,
  plus_grp = 15,
  lw_tol = 0.1,
  lw_maxiter = 1000,
  weight_scale = 1000,
  by_month = FALSE,
  digits = 5
)
}
\arguments{
\item{d}{A data frame as returned by \code{\link[=extract_commercial_samples_sog]{extract_commercial_samples_sog()}} or
\code{\link[=extract_survey_samples_sog]{extract_survey_samples_sog()}}}

\item{min_date}{Earliest date to include}

\item{raw_counts}{Logical. If \code{TRUE}, return raw, unweighted age
proportions. If \code{FALSE}, return the age proportions weighted by sample
and catch weights}

\item{plus_grp}{Age plus group for maximum grouping}

\item{lw_tol}{See \code{\link[=fit_lw]{fit_lw()}}}

\item{lw_maxiter}{See \code{\link[=fit_lw]{fit_lw()}}}

\item{weight_scale}{A value to divide the weights by}

\item{by_month}{Logical. If \code{TRUE}, return a data frame with a
\code{month} column in addition to a \code{year} column}

\item{digits}{The number of decimal places to round the values to}
}
\value{
The age proportion data frame, invisibly
}
\description{
Create an age proportion CSV data file from the sample data
}
\details{
A CSV file will be written, and the data frame will also be returned
invisibly. The data frame will contain the year, number of fish by year
in the data, and the number of samples (hauls/trips) in the data.
Shoreside uses trips as a sampling unit and the other two use hauls.
}
