% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create-month-catch-csv-file.R
\name{create_month_catch_csv_file}
\alias{create_month_catch_csv_file}
\title{Create a CSV file with catch by year and month}
\usage{
create_month_catch_csv_file(df, catch_scale = 1000, digits = 3)
}
\arguments{
\item{df}{A data.frame as returned by \code{\link[=extract_catch]{extract_catch()}}}

\item{catch_scale}{A value to divide the catch by prior to writing to
the file}

\item{digits}{The number of decimal points to report on the catches in
the csv files}
}
\value{
Nothing, creates three CSV files
}
\description{
Create a CSV file with catch by year and month
}
