% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create-year-catch-csv-file.R
\name{create_year_catch_csv_file}
\alias{create_year_catch_csv_file}
\title{Calculate the catch by year}
\usage{
create_year_catch_csv_file(
  df,
  catch_scale = 1000,
  write_file = TRUE,
  digits = 2
)
}
\arguments{
\item{df}{A data.frame as returned by \code{\link[=extract_catch_sog]{extract_catch_sog()}}}

\item{catch_scale}{A value to divide the catch by prior to writing to
the file}

\item{write_file}{If \code{TRUE}, write the output to the data file given by
the package data variable \code{catch_by_year_fn}. If \code{FALSE}, return the
data frame}

\item{digits}{The number of decimal places to include in the table output}
}
\value{
If \code{write_file} is \code{TRUE}, nothing. If \code{write_file} is \code{FALSE},
a data frame containing the catch by year (rows) and fishery type (columns)
}
\description{
Calculate the catch by year
}
