% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depth_sql_fn}
\alias{depth_sql_fn}
\title{Name of file containing the SQL code to extract Canadian fishery gear
depths and bottom depths}
\format{
A character string
}
\usage{
depth_sql_fn
}
\description{
Name of file containing the SQL code to extract Canadian fishery gear
depths and bottom depths
}
\keyword{datasets}
