% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract-catch.R
\name{extract_catch}
\alias{extract_catch}
\title{Merge DMP and LOGS catch data and extract into a data frame}
\usage{
extract_catch(lst)
}
\arguments{
\item{lst}{a list of two data frames as returned by
\code{\link[=load_catch_data]{load_catch_data()}}}
}
\value{
A list of three data frames, one for each gear: Freezer trawlers,
Shoreside, and Joint Venture
}
\description{
Merge DMP and LOGS catch data and extract into a data frame
}
