% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-catch-sog.R
\name{extract_catch_sog}
\alias{extract_catch_sog}
\title{Extract the Strait of Georgia catch by year for the SS3 input file}
\usage{
extract_catch_sog(ct)
}
\arguments{
\item{ct}{The output of \code{\link[gfdata:get_data]{gfdata::get_catch()}}}
}
\value{
A data.frame of the catch by year for hake in the SoG
}
\description{
Extract the Strait of Georgia catch by year for the SS3 input file
}
\details{
Cannot use \code{trp_type} here as you would expect. The trip types for the SoG
vary from the GULF hake (as expected) to "Option A Hake Shoreside" to \code{NA} so
we just extract by area and gear type to get midwater caught hake in
area 4B (excluding the Strait of Juan De Fuca). Landings and Discards are
added together.
}
