% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-commercial-samples-sog.R
\name{extract_commercial_samples_sog}
\alias{extract_commercial_samples_sog}
\title{Extract the Strait of Georgia commercial samples by year for the SS3
input file}
\usage{
extract_commercial_samples_sog(d)
}
\arguments{
\item{d}{The output of \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}
}
\value{
A data.frame of the commercial samples by year for hake in the SoG
}
\description{
Extract the Strait of Georgia commercial samples by year for the SS3
input file
}
\details{
Cannot use \code{trp_type} here as you would expect. The trip types for the SoG
vary from the GULF hake (as expected) to "Option A Hake Shoreside" to \code{NA} so
we just extract by area and gear type to get midwater caught hake in
area 4B (excluding the Strait of Juan De Fuca). Landings and Discards are
added together.
}
