% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-survey-samples-sog.R
\name{extract_survey_samples_sog}
\alias{extract_survey_samples_sog}
\title{Extract the Strait of Georgia survey samples by year for the SS3
input file}
\usage{
extract_survey_samples_sog(d)
}
\arguments{
\item{d}{The output of \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}}}
}
\value{
A data.frame of the survey samples by year for hake in the SoG
}
\description{
Extract the Strait of Georgia survey samples by year for the SS3
input file
}
