% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catchability-density.R
\name{plot_catchability_density}
\alias{plot_catchability_density}
\title{Plot the density histogram for the posteriors for catchability for
the acoustic age-2+ survey}
\usage{
plot_catchability_density(
  model,
  num_bins = 30,
  x_breaks = seq(0, 1.2, 0.1),
  y_breaks = seq(0, 3, 0.5),
  bar_outline_color = "black",
  bar_fill = main_fill,
  bar_alpha = main_alpha,
  line_color = "green",
  line_type = "solid",
  line_width = 1
)
}
\arguments{
\item{model}{The base model}

\item{num_bins}{The number of bins to use for the histogram created using
\code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}}

\item{x_breaks}{A vector of values to show on the x-axis}

\item{y_breaks}{A vector of values to show on the y-axis}

\item{bar_outline_color}{The color of the outline of each bar in the histogram}

\item{bar_fill}{The color of the fill for each bar in the histogram}

\item{bar_alpha}{The amount of transparency for each bar in the histogram}

\item{line_color}{The line color to use for the median line}

\item{line_type}{The line type to use for the median line}

\item{line_width}{The line width to use for the median line}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the density histogram for the posteriors for catchability for
the acoustic age-2+ survey
}
