% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_age_counts_fn}
\alias{raw_age_counts_fn}
\title{The filename for the raw age counts by year data file}
\format{
A character string
}
\usage{
raw_age_counts_fn
}
\description{
The filename for the raw age counts by year data file
}
\keyword{datasets}
